review: guard against pending exception; move tests to process-stdio.test.ts
- Skip the JS emit path when a JS exception is already pending so the
  C++ handler's exception scopes can't swallow the user's original throw.
  The pipe stays broken so the next console.* call retries.
- Move the #7251 tests into test/js/node/process/process-stdio.test.ts
  (this is a Node-compat gap, not a regression) and make them concurrent.
